fix: 사용자 도메인 Clean + Layered Architecture 적용
