+++ src/main/java/kr/hhplus/be/server/application/user/UserService.java
@@ -1,5 +1,7 @@
-package kr.hhplus.be.server.user;
+package kr.hhplus.be.server.application.user;
 
+import kr.hhplus.be.server.domain.user.User;
+import kr.hhplus.be.server.interfaces.user.UserResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -7,7 +9,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class UserService {
 
-    public User chargeBalance(long userId, long amount) {
+    public UserResponse.User chargeBalance(long userId, long amount) {
 
         // 사용자 식별자 유효성 체크
 
@@ -20,16 +22,16 @@ public class UserService {
 
         // 잔액 충전
 
-        return new User(1L, "홍길동", 1000);
+        return new UserResponse.User(1L, "홍길동", 1000);
     }
 
-    public User getBalance(long userId) {
+    public UserResponse.User getBalance(long userId) {
 
         // 사용자 식별자 유효성 체크
 
         // 사용자 식별자로 사용자의 잔액 조회
 
-        return new User(1L, "홍길동", 1000);
+        return new UserResponse.User(1L, "홍길동", 1000);
     }
 
 }
