+++ src/main/java/kr/hhplus/be/server/domain/user/User.java
@@ -1,0 +1,11 @@
+package kr.hhplus.be.server.domain.user;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+public record User(
+    long id,
+    String name,
+    int balance
+) {
+
+}
