+++ src/main/java/kr/hhplus/be/server/interfaces/user/UserController.java
@@ -1,5 +1,7 @@
-package kr.hhplus.be.server.user;
+package kr.hhplus.be.server.interfaces.user;
 
+import kr.hhplus.be.server.application.user.UserService;
+import kr.hhplus.be.server.domain.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,13 +18,13 @@ public class UserController implements UserApi {
 
     @Override
     @PostMapping("/{userId}/charge")
-    public User chargeBalance(@PathVariable long userId, long amount) {
+    public UserResponse.User chargeBalance(@PathVariable long userId, long amount) {
         return userService.chargeBalance(userId, amount);
     }
 
     @Override
     @GetMapping("/{userId}/balance")
-    public User getBalance(@PathVariable long userId) {
+    public UserResponse.User getBalance(@PathVariable long userId) {
         return userService.getBalance(userId);
     }
